Changed: dispatch()
dispatch is either called

    dispatch($key, KEY1 => $sub1, KEY2 => $sub2);

or

    my $func = dispatch({KEY1 => $sub1, KEY2 => $sub2});
    for my $x ( @array ) {
        $func->($x->{key}, $x);
    }

Either the key with a mapping to a subref is passed. This is usually always
done exactly where it needs to be done. Because very very likely we need
to access scoped variables.

For optimization, we can pass a hash-ref with the mappings. Then it returns
a $func that can be used. In this function we just pass the key that
needs to be dispatched. Any amount of extra arguments can be passed. Usually
this will be the item itself.

The second version has the benefit that a hash does not be created every
time. But i consider to just use Array::dispatch instead.

Still don't know if i remove the case with just one argument. Technically
it makes sense to just use Array::dispatch instead.