Removed Array::dispatch
After a lot of adding/removing i stick with dispatch() in two variants.
It either can be called with a key and expects a list of all the dispatches
based on that key. This is usually how you use it somewehre directly
in the code. For example in a loop.

Otherwise there is a special two argument version that creates a new function.
dispatch() with two arguments expects a function to generate a key from
the passed in value, and then picks the correct function from a dispatch
table.

The second version can be used in Array::iter, Array::map and maybe
in a lot of other cases. Some more examples have to be discovered yet.

The second version makes Array::dispatch completely obsolete.