Changed: Sq::Test::dies
dies() is now a full test function on its own. Previously dies() basically
just catched the exception message and returned it. You then had to use
like() to test for an error. I used it this way because i bacially
imported the API from Test2.

But this API isn't really good. In the case when code doesn't throw an
exception it just returned an undef. like() then must check for undef
values. The problem is that like() is a generall function to just test
if a string matches any regex. Assuming that an undef was a not thrown
exception isn't possible anymore. That's a crap API.

So i changed dies() to be a full test function on its own. You write
some code, provide the error message you expect with a regex and dies
checks if an exception is thrown and the message matches. If either
is not correct it gives you useful messages telling you what went wrong.

Either that the message didn't match or the code didn't throw an exception.