Improved: Sq::Fs::read_text
Before opening a file i test if it is a file, this way directories
are skipped.

Previously the PerlIO Layer did the UTF-8 encoding, but when opening a file
that didn't contain valid utf-8, because for example it is a binary file,
then the code printed a lot of warnings and as a user of the code, you
didn't had a clue which file caused the problem.

Because of that i now do the utf8 decoding manually and check the return
value. If the decoding fails then a warning is printed which file causes
the problem and reading from that file is aborted.