Changed loading Order
I now first load Dump, Equality and Copy, So those functions can be
easier used in the collections. I now also can create a reference
to them at the start of those modules instead of fully qualify the
functions all the time. Usually this is also a little bit faster.