Fmt::table: Now also supported sequence
A sequence is transformed into array before calling function. So
a sequence can now be passed without explicitly calling ->to_array.

This operation is not lazy, but printing a table anyway cannot be lazy.
The column size of every line must be calculated beforehand.

Theoretically could be lazy if an exact column-size for every field is
provided. But only then this could be lazy at all.