Changed: Array/Seq:combine: combine now expects a sub-ref to determine the key
This is a lot more flexible. Also is a lot more in common how it works
with any other function so far. When a unique from a hash should be
directly selected, you can just use `key 'id'`.

Otherwise any other sub-ref can be provided to calculate uniqueness
however it is needed. Sometimes not every hash always contains it's
unique key and it must be somehow calculated. This small little change
allows todo that.