Enhanced: Array::fill2d
Because Perl Arrays are mutable by default, in a lot of cased where a
default value is expected, for example for initialization of a value,
a subroutine is expected. This is important because without a sub-ref
only one single value is created and shared. This was also the case
for fill2d. Not providing a sub-ref meant one array is created and this
array would be used for every field that is filled.

But because copy() is now implemented i can differentiate between
both cases. Now i just can support that a user still can provide a
function to provide an initial value, or just pass any mutable
data-structure. I just can copy the value.

To help in the creation, the init() function is introduced. Wherever
it is needed you just can call `init($f_or_value, @args)` and either
`$f_or_value->(@args)` is executed and the result is taken, or a copy
of `$f_or_value` is made.

Now someone can write.

    Array::fill2d($array, []);

and the array is filled with array-refs, and every array-ref is a new
array. It is just DWIM.

More functions will be changed to support this from now on.