Type: Added t_and
Didn't implemented it, because most functions are written that they support
many types by default. So t_and is bacially already implemented in most
functions. This was also done because of performance and how to write
types, it is annoying to always write t_and when more than one check would
be needed. Still t_and can be helpful in some cases to combine some checks
into more complex type-checks.