Added: t_maybe() | Changed: t_keys
t_keys usually expected that all keys must exists on a hash. This
is not a requirement anymore. Still this doesn't change much. As when
a key doesn't exists, it is undef, and undef is passed to the next
check, and usually all checks don't allow undefs.

Also ensured that all type-checks so far, does an undef check. Missed
some functions that did that.

This change is made because of t_maybe() to work together with t_keys.
t_maybe() is a type-check that check if a value is defined against
a type, or if the value is not defined, it is still valid.

So a value can be not present, or when it is present, it must be valid.