Refactor Tests - t_valid -> t_run
Rewritten all the tests to use t_run instead of t_valid. t_valid only
returned a bool if type was successful or not. t_run just returns
the Result and ok() expect an Ok() value.

For tests this is a lot better. Because in an Err() case a useful error
message is dumped, instead of just returning false.