Removed Carp::croak calls and checks in Sq::dispatch
This is the whole feature of Sq::Signature. By default i started with
dispatch() and checked all invalid cases and did throw errors in
the not allowed cases. But this means dispatch() is full of error-handling.

This error-handling costs performance, as it always happens. You cannot
deactivate it. Sure the error-handling is there that someone correctly
calls dispatch() and early errors are found.

But this is the whole purpose of Sq::Signature. Now the invalid cases
are checked by Signature. The advantage is that when someone
deactivates the signature, all the runtime-checks disappear. Making
dispatch() faster when signature is off.