Sq->fmt->table: header with array of hashes is now optional
The header doesn't need to be specified anymore. But when it is not
specified the header will be automatically calculated by fetching all
keys from every hash and sorting them.

When printing just array of hashes as a table and no specific column order
is expected then this can be helpful.