fmt: table: A little bit more specific to the type
It basically only supports AoA or AoAoA to a depth of 3 at the moment.
In the future i guess i will just create a flexible grid system to
describe a layout.

A flexible grid system also can be easily mapped to all kind of other
layouting, like generating HTML or generating an image and so on.