Signature: Replaced sigt() with new version
The new version now expects multiple IN => OUT to be passed instead
of just supporting a single IN => OUT. This works a lot better for
functions with multiple different IN => OUT, and the IN => OUT also
must match each other.

With the previous version this was not possible.

The new version is basically API backwards compatible. All older calls
should work the same. The new version is just named sigt() again. Also
sig() just calls sigt() and everything is still working.

With experimentation i saw the the correct IN => OUT was matched, but
still must add some tests for that too.