Added no-checks version
While some additional checks can be passed to some functions, those
additional checks are optional. In the case no checks are passed the
returned lambda can be cached. Also the code itself can be shortened
a little bit because we don't need to loop through checks.

This change can improve runtime and reduce memory consumption. Especially
when someone don't explicitly cache the results into variables
at the beginning of the signature.

On the other hand it makes initialization a little longer because
the t_* functions now has additional conditionals. But this shouldn't be
much. Lowering the runtime overhead should be more important.