Array/Seq: fold: Now both uses init()
Theoretically it always maked sense to better use a sub-ref here. But
i could get away with it because most Array/Hash functions return new
data-structure, so it didn't mattered much even when you start with
a mutable variable. In maybe 98% of all cases it would work correctly.

But also that i explicitly expect that someone must return a value, and
by default all mutable methods return nothing lead to a case where it is
nearly impossible to use it wrong.

In the tests i use push() for mutation and still must return the value
itself to work. This will not change. Still using init() is what might be
logical for such a function. It makes the API rememberable.

fold needs an initial state and the init can be a value or sub-ref.

But it also has the nice additions that someone now can explcitily just
pass [] or {} and still get a blessed version and it will still be an
empty copy even if someone uses mutation inside fold().