Fixed: Floating Point Comparision
By default you cannot really compare floats. Comparing floats is basically
always a mistake. Because of floating point inaccuray you cannot even
say that "0.1 * 10" is the same as "1". It could be "0.999999999999987"
or stuff like that.

I now decided to fix that. What is the point of an equal() that doesn't
work when it has floating point in it?

The typical way of comparing two floats is by subtracting them, and then
check if the two numbers are beyond a specific tolerance. I picked

1e-08 as the default tolerance. This is also what Test2 uses as the
default.

This can have a potential impact on the whole system. Because every number
comparision goes through it. But running test-suite doesn't really made
much of a difference at all.

But i also don't have much number comparision in my tests.

But even when it is slower. What is the point when code is fast, but it
cannot compare floats "correctly"?

The Tolarenace can be changed by Setting the global variable. For example:

$Sq::Equality::FloatTolerance = 0.0001;