record: Now expects exact arguments
The idea of an Record is somehow choosen from F#. Technically a record
in F# is the same as a Hash. Only statically typed. But a Record always
must have every field initialized. It's not allowed to skip fields.

Optional fields are implemented in F# by expecting an Option field
instead.

The record() function is not only an easy way to turn some kind of list
into a hash. Expecting the exact amount of values gets rid of some
errornous code where you either pass the wrong amount of arguments, or
didn't pass all fields to `record()`.

This was also the case with the included example. When writing a wrapper
around stat() i forgot a field and the returned hash was somehow incorrect.