Added: Sys->sys->capture
A way to run a shell command and get STDOUT/STDERR from program, wraps
it in an Result.

Used Sq->io->youtube with it. It seems as soon there is a WARNING yt-dlp
automatically returns error code 1, even when everything is succesfull.
So i just print STDERR in that case and still just use the output.

Maybe later i can fix this better, but i have no patience todo that now.