Changed implementation
Now i just pass the $canvas itself to the lambda. So just the lambda
itself it basically lazy created. This is a lot more flexible is
i can create more basic functions with some better performance.

The offset feature is now implemented as part of $canvas, it holds
the current offset and can be changed with addOffset.