Previously cwrite
The cwrite does an "intelligent" clipping and calculates what needs
to be written. Theoretically with big strings that are outside
the canvas, this should be faster. As only what needs to be written
is actually written. In practice this seems to be slower.

The easier setChar() who just skips every char and checks every position
explicitly is faster. One reason is surely because i usually don't
provide strings and positions outside the canvas.

Still leave it here commented. So i can check the other impelemtation
from time to time.