Canvas: Changed implementation
Implementation changed again. First i Used an char array. It worked
but it seemed slow to me, maybe too much memory. So i switched to
using a string. As strings are usually arrays anyway.

But after all changes. Performance was the same. Iterating a string
is actually not really efficent. In Perl there are not arrays. Only
way to acces individual characters is by either splitting or using
substr().

In benchmark it turned out that a char array is the fastest. But it takes
a lot more memory. Instead of for example 22K for a total screen, it is
about 600K memory. But it is still around 4 times faster.

Still changed again to array of strings. After some code writing
i think this is the best. It is not the fastest. A little bit faster compared
to just using a single string.

But all kind of line access is a lot easier to code. Also a array of strings
allows for easier feature like expanding in width/height. Before only
expanding in height was easy.