Canvas: re-added set_char()
I started to soon thinking about performance and optimize stuff. So
i created setChar at the beginning that could write a whole string.
But when features are added like special character handling and so on
it becomes more complicated.

It's good to always have the most primitive operations and just use them.
For example using write_str() with all kind of character handling
when i really just need to set a single char in a line drawing algorithm
then using write_str() seems unapropiate.

But also other code like write_str() and put() are more complicated,
because all of them have to re-implement the offset handling again.

I leave write_str() and put() as they are, because the inlined code might
be faster instead of code that always calls set_char(). But when starting
development it is usually good to start with primitives and built on
top of them. This is really a good concept i learned from functional
programming and it works very well.

Optimization maybe could be added later when something seems feature
complete.