Canvas: Changed offset to clip
I now start to enhance offset. Offset itself is less useful. What i
really need is to specify an area where draw comments will end up.
For this i not only need an offset, but also to be able to specify
width & height.

Because how text is working it also makes sense to just limit something
to width, and let stuff grow in height when characters are put().
It basically emulates how a terminal works.

For a typical real image this wouldn't make sense, but here i am dealing
with text creation for display on console.

Currently only changed offset to clip and provided functions to change
the clip or read values, so at the moment it is backwards compatible
but the new functionality is not yet implemented.