Removed Offset/Clip
In the last commit i wanted to introduce clipping, but instead removed
it completely, including the ability to set an offset. Still changed
everything that no feature so far is lost. Here is the idea and reason.

So far i had the ability to set an Offset, this was nice as i easily
could change drawing operations and for example let them write into
another area. But just an offset was not enough.

Sometime i want to be able to specify an area and all write operations
should restrict just to that area. With a simple offset this was not
enough. I also had to cut off thing that go outside the restricted
area.

Then i also wanted to be able to just set a width, and when put() is
called the put just restricts to the defined width and offset, but
automatically grows. The idea with clipping would add those features.

The problem was. It was not really needed. There already was a way to
have those things. Just create a new canvas, do you set() and put()
operation and then just merge() or clip() the new canvas into the old
one.

One reason for the clipping, so i thought, was "better performance".
Because instead of creating a new canvas it would make every
operation on a mutable canvas and changing a canvas. The problem
with this approach is the implementation. Because when clipping
is added then this feature must be added to the most basic function
set_char() and put_char().

The result would be a basic function that now not only write a character
but also does offset handling and clipping. And it must handle three
different cases. Changes are that this does not improve performance,
it makes everything worse. And all of that added complexity for a feature
that is already achivable?

So i stopped that idea. Not only did i stop it. I thought of also removing
the offset making it more simpler. It took me quite some approaches
until i finally came to a solution that worked without too much changes.
I guess i will improve the API, especially the combinator one, but
so far it is working again, it has the same features and everything with
the removal of the offset ability.

Removing the offset ability was somehow tricky because creating another
canvas and merging it was not always the same. Because every canvas
has a default character merging another canvas always would reset
every character. But sometimes i just wanted to be able to save some
write operation into a chunk and just insert those exact characters
that are overwritten, not the whole area.

I solved that problem by introducing an "Empty Canvas". This is basically
also the idea of Monoid. Always have an "Unit" value that does nothing.

Like "5+0" and "0" is the unit value for "+" that doesn't change anything.

You can think of it as a way to create transparent canvases, and now you
just can combine them. Still i must do same more changes and so some
more tests for that feature, maybe some further API changes, but so
far everything is working and I like that approach better then
putting more and more features into set_char() & put_char().

Removing the offset also made some stuff easier. For example before i
had two canvases and both could have an offset set. And when merging you
could specify an additional offset. It worked, but it was hard to
understand what the hell was even created. Now this is a lot easier.

I solved the problem of restrictiong and growing by providing a merge()
and clip() function.

merge() basically places a canvas into another canvas with an offset and
grows the height when needed. clip() does the same but without growing.