Added: mean(), mean_by()
Those functions are mostly the same as average() and average_by(), it
just uses a different algorithm to calculate the the average.

The average functions use a simple sum/count algorithm. The benefit of
it is that it is faster. it also uses List::Util::sum0 a C implementation
to calculate the sum. The disadvantage of this approach is that it
is less numerical stable.

When summing a lot of values. The sum can become very big (at least when it
does not contain negative numbers). Or otherwise grow big in the negative.
Because of floating point there is a limit on how accurate the sum can
be.

The mean() functions on the other hand computes the average with an iterative
approach. Each new value basically is already the average so far. Each new
value will only slightly shift the average and corrects it a little bit
and makes it bigger, smaller or the same.

This results into better numerical stability. But also has the disadvantage
that it involes a lot more computation. It also is written in purely Perl
and has no C implementation.

Choose mean() or mean_by() when you have a lot of numbers and/or your numbers
become very big/low and you need an accurate value as much as possible.
Otherwise just pick average()