Dump: Deactivate color with dumps()
Color by default emits Shell sequence for the terminal for color. It
is best when we want to print to the console by default. As dump()
print to the STDERR by default, and assumes it the console, it is
active by default.

But when using dumps() and we want a string, then usually this is for
other reasons. Like serialization (even when this is not the purpose of
Dump so far and has a bug), or maybe printig to a log-file and so on.

In this cases we rarely want the string to contain shell escape sequences
for colors.