Hash,Array,Seq: Added get_init(), fix replicate
Added a get_init() function that allows to get a value from Hash/Array,
and when it doesn't exists, it creates a default value for that index
or key.

get_init() expects $init, so a user can specify a value that is copied
or a function that is executed.

During tests i also fixed replicate() that now does copies of
data-structures that are passed.

Seq::get_init() theoretically doesn't make much sense, because a sequence
lazy evaluates all data and creates them on-demand. At least when a
sequence doesn't use mutable state.

It anyway isn't a good function to be called because it must iterate
through the sequence to find the value. But it still exists because
of API compatibility with Array.

It is still better as when someone calls ->to_array->get_init() only to
maybe fetch a single element, and then throws away the array anyway.