Removed Sq::List from AutoLoad
Sq::List is not updated since ages. It has a lot of less functionality
compared to Seq/Array. It is missing most functions, and most are not
api compatible. Also never added signatures for List.

Even Seq/Array are still not fully compatible as i wish and they are
missing functions in the one or other package.

Still, even if List would be fully compatible, i wouldn't suggest to use
List. Because it doesn't really have much advantage over an Array.
Most Array functions anyway return a new array. So in that sense
they are immutable anyway. But creation of Array, iteration and so on
are probably magnitudes faster than using a List.

List still can be loaded manually if wished.