Changed: seq {} -> seq()
There are reasons why it was like before, and why i changed it. The syntax
was from F#. There you could write seq { ... } and inside it you had
"normal code". With "yield" you could emit one item after another.
In F# this is a special construct called "Computation Expression"
that is specially supported by the compiler.

Perl has no such feature, but i still could implement seq { } and made it
a normal perl subroutine that executes code once, and uses the result
of it as normal values.

But, this is still not an iterator, for this you had to use something like
Seq->from_sub. So seq {} was primarily used as a quick way to just create
a sequence out of some values.

The problem here was that it is not consistent with array() or hash()
that are also just functions that expects normal values. So i just
changed everything. Now seq() is exactly like array() or hash()
and just gets some values to create the data-structure.

When someone wishes to have some code that in the end creates a sequence,
you just can use assign {} and return seq() at the end.