Locking keys
This helps with accessing not availble keys. It then throws an exception.
This happened already to me. Used ->{path} instead of ->{PATH} somewhere
and got an error message about undef not implementing a method. The
wrong error message at nearly the wrong place.

Sometimes the behaviour is used that not defined keys just return undef.
We then just check for defined. Instead of doing this either check
with "exists" and then extract, or just use "Hash::get". This returns
an optional even for the case of an undef.

    ->get("path")->match(Some => ..., None => ...)

the above also would have not caused a crashing program and maybe helped
in a better error message. But i was lazy ...