Type: Added t_tuplen
A tuple where only the minimum amount of elements are needed. This is
usually helpful in functions with default arguments. Currently i
needed to write something like

t_or(
    t_tuple(t_int)
    t_tuple(t_int, t_int)
)

to specify that it either can be one int or two int. Now it can be written
as

t_tuplen(1, t_int, t_int)