Added TODO: parse hex numbers
Parsing hex numbers is only supported up to 32bit. Even if it still works
for bigger numbers. But it prints a warning. I have to go through
documentation why this is the case.

If hex only can parse bigger hex numbers on 64bit then i am fine with
that restriction. Most Linux distribution not even have a 32bit version
anymore?